requirements.txt
tqdm>=4.66.1
numpy>=1.26.0
torch>=2.2.0
torchvision>=0.17.0
easydict>=1.10
pandas>=2.2.0
jupyter>=1.0.0
pyyaml>=6.0.1
ipykernel>=6.29.0
scipy>=1.12.0
nltk>=3.8.1
tensorboard>=2.15.0
ftfy>=6.1.3
regex>=2023.12.25
flask>=2.3.0
pillow>=10.0.0